+++ UpdateChecker/ModVersion.cs
@@ -1,0 +1,11 @@
+﻿using Semver;
+
+namespace UpdateChecker
+{
+    public class ModVersion
+    {
+        public string modversion;
+        public SemVersion SemVersion;
+        public int ApprovalStatus;
+    }
+}