+++ WorldPredownload/DownloadManager/WorldDownloadManager.cs
@@ -141,23 +141,31 @@ namespace WorldPredownload.DownloadManager
             // ReSharper disable once ConditionIsAlwaysTrueOrFalse
             webClient?.Dispose();
             webClient = new WebClient();
-            webClient.Headers.Add("user-agent", ModSettings.downloadUserAgent);
+            #if DEBUG
+            MelonLogger.Msg($"Using UserAgent: {ModSettings.downloadUserAgent}");
+            #endif
+            webClient.Headers.Add(HttpRequestHeader.UserAgent, ModSettings.downloadUserAgent);
             webClient.DownloadProgressChanged += progress;
             webClient.DownloadFileCompleted += compete;
 
             var cachePath = CacheManager.GetCache().path;
-            var assetHash = CacheManager.ComputeAssetHash(apiWorld.id);
+            var assetHash = CacheManager.ComputeAssetHash(apiWorld.assetUrl);
             var dir = Path.Combine(cachePath, assetHash);
             if (!Directory.Exists(dir)) Directory.CreateDirectory(dir);
-            var assetVersionDir = Path.Combine(dir,
-                "000000000000000000000000" + CacheManager.ComputeVersionString(apiWorld.version));
+            var assetVersionDir = Path.Combine(dir, "000000000000000000000000" + CacheManager.ComputeVersionString(apiWorld.assetUrl));
             if (!Directory.Exists(assetVersionDir)) Directory.CreateDirectory(assetVersionDir);
 
             var fileName = Path.Combine(assetVersionDir, "__data");
+            #if DEBUG
+            MelonLogger.Msg($"Calculated Directory: {assetVersionDir}");
+            #endif
             MelonLogger.Msg($"Starting world download for: {apiWorld.name}");
             file = fileName;
 
-
+            #if DEBUG
+            MelonLogger.Msg($"AssetUrl: {apiWorld.assetUrl}");
+            #endif
+            
             webClient.DownloadFileAsync(new Uri(apiWorld.assetUrl), fileName);
         }
     }