+++ WorldPredownload/Helpers/Utilities.cs
@@ -1,6 +1,7 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using Il2CppSystem;
 using MelonLoader;
 using Transmtn.DTO.Notifications;
@@ -18,12 +19,16 @@ namespace WorldPredownload.Helpers
     [SuppressMessage("ReSharper", "HeuristicUnreachableCode")]
     public static class Utilities
     {
+        private static readonly Regex fileIdRegex = new("/file_[0-9A-Za-z-]+/", RegexOptions.Compiled);
+
+        private static readonly Regex fileVersionRegex = new("(?:\\/file_[0-9A-Za-z-]+\\/)([0-9]+)", RegexOptions.Compiled);
+
         public static void AdvancedInvitesHandleInvite(Notification notification)
         {
 #if DEBUG
             try
             {
-                GetAdvancedInvitesInviteDelegate(notification);
+                //GetAdvancedInvitesInviteDelegate(notification);
             }
             catch (Exception e)
             {
@@ -135,6 +140,16 @@ namespace WorldPredownload.Helpers
         public static PageUserInfo GetUserInfo()
         {
             return GameObject.Find(Constants.PATH_TO_USERINFO).GetComponent<PageUserInfo>();
+        }
+        
+        public static string ExtractFileId(string txt)
+        {
+            return fileIdRegex.Match(txt).Groups[1].Value;
+        }
+
+        public static string ExtractFileVersion(string txt)
+        {
+            return fileVersionRegex.Match(txt).Groups[1].Value;
         }
     }
 }