+++ WorldPredownload/Main.cs
@@ -3,9 +3,7 @@ using UIExpansionKit.API;
 using WorldPredownload.Helpers;
 using WorldPredownload.UI;
 
-[assembly:
-    MelonInfo(typeof(WorldPredownload.WorldPredownload), "WorldPredownload", "1.5.3", "gompo",
-        "https://github.com/gompocp/VRChatMods/releases/")]
+[assembly: MelonInfo(typeof(WorldPredownload.WorldPredownload), "WorldPredownload", "1.6.2", "gompo", "https://github.com/gompocp/VRChatMods/releases/")]
 [assembly: MelonGame("VRChat", "VRChat")]
 [assembly: VerifyLoaderVersion(0, 4, 3, true)]