+++ WorldPredownload/UI/WorldButton.cs
@@ -102,13 +102,23 @@ namespace WorldPredownload.UI
             }
             else
             {
-                if (CacheManager.HasDownloadedWorld(world.id, world.version))
+                if (CacheManager.HasDownloadedWorld(world.assetUrl))
                     Button.SetText(Constants.BUTTON_ALREADY_DOWNLOADED_TEXT);
                 else
                     Button.SetText(Constants.BUTTON_IDLE_TEXT);
             }
         }
 
+        public static void UpdateTextDownloadStopped()
+        {
+            var worldInfo = GetWorldInfo();
+            if (worldInfo == null || worldInfo.field_Private_ApiWorld_0?.id == null) return;
+            if (CacheManager.HasDownloadedWorld(worldInfo.field_Private_ApiWorld_0.assetUrl))
+                Button.SetText(Constants.BUTTON_ALREADY_DOWNLOADED_TEXT);
+            else
+                Button.SetText(Constants.BUTTON_IDLE_TEXT);
+            CanChangeText = true;
+        }
 
         private static PageWorldInfo GetWorldInfo()
         {